+++ NuGet.Services.Staging.sln
@@ -42,6 +42,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "client", "client", "{FFC42A
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NuGet.CommandLine.StagingExtensions", "src\client\NuGet.CommandLine.StagingExtensions\NuGet.CommandLine.StagingExtensions.csproj", "{3B7D7E32-327A-4DF2-B7DD-0EB2B11FD761}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "client", "client", "{EF3E219F-8E86-47D6-8B5F-ECA217070367}"
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "NuGet.CommandLine.Staging.UnitTests", "test\client\NuGet.CommandLine.Staging.UnitTests\NuGet.CommandLine.Staging.UnitTests.xproj", "{1646DD7B-E86C-43D9-9DC0-5460F2A699BC}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -104,6 +108,10 @@ Global
 		{3B7D7E32-327A-4DF2-B7DD-0EB2B11FD761}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3B7D7E32-327A-4DF2-B7DD-0EB2B11FD761}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3B7D7E32-327A-4DF2-B7DD-0EB2B11FD761}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1646DD7B-E86C-43D9-9DC0-5460F2A699BC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1646DD7B-E86C-43D9-9DC0-5460F2A699BC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1646DD7B-E86C-43D9-9DC0-5460F2A699BC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1646DD7B-E86C-43D9-9DC0-5460F2A699BC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -124,5 +132,7 @@ Global
 		{49FD2E6D-594F-4998-AEAF-45BC26DB41BB} = {FFC42A02-91E3-4BF0-9D65-EAF524E48881}
 		{FFC42A02-91E3-4BF0-9D65-EAF524E48881} = {2FF49FA1-E459-43C4-A86D-8EB40DF7FCC2}
 		{3B7D7E32-327A-4DF2-B7DD-0EB2B11FD761} = {FFC42A02-91E3-4BF0-9D65-EAF524E48881}
+		{EF3E219F-8E86-47D6-8B5F-ECA217070367} = {01B6E97F-7B86-4364-AFD3-81881C199117}
+		{1646DD7B-E86C-43D9-9DC0-5460F2A699BC} = {EF3E219F-8E86-47D6-8B5F-ECA217070367}
 	EndGlobalSection
 EndGlobal
