+++ src/client/NuGet.Client.Staging/Views/MembershipView.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace NuGet.Client.Staging
+{
+    public class MembershipView
+    {
+        public string Name { get; set; }
+        public string MembershipType { get; set; }
+    }
+}
