+++ src/client/NuGet.Client.Staging/Views/StageCommitProgressView.cs
@@ -1,0 +1,16 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System.Collections.Generic;
+
+namespace NuGet.Client.Staging
+{
+    public class StageCommitProgressView : StageView
+    {
+        public string CommitStatus { get; set; }
+
+        public List<PackageCommitProgressView> PackageProgressList { get; set; }
+
+        public string ErrorMessage { get; set; }
+    }
+}
