+++ src/client/NuGet.Client.Staging/Views/StageListView.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace NuGet.Client.Staging
+{
+    public class StageListView : StageView
+    {
+        public string MembershipType { get; set; }
+    }
+}
