+++ src/client/NuGet.Client.Staging/project.json
@@ -7,7 +7,6 @@
         "NuGet.Protocol.Core.v3": "3.5.0-beta-final"
     },
     "frameworks": {
-        "net45": {
-        }
+        "net45": {}
     }
 }
